pyproject.toml
[tool.poetry]
name = "imserv"
version = "0.1.11"
description = "Spin an image server, store images from Clipboard in single place, and prevent duplication."
authors = ["Pacharapol Withayasakpunt <patarapolw@gmail.com>"]
repository = 'https://github.com/patarapolw/imserv'
homepage = 'https://github.com/patarapolw/imserv'
readme = 'README.md'
license = "MIT"

[tool.poetry.dependencies]
python = "*"
nonrepeat = "^0.1.1"
flask = "^1.0"
python-slugify = "^1.2"
imagehash = "^4.0"
psycopg2-binary = "^2.7"
pillow = "^5.2"
tqdm = "^4.26"
watchdog = "^0.9.0"
send2trash = {git = "https://github.com/acheronfail/send2trash.git", branch = "fix/trash-put-back-macos"}
peewee = "^3.7"
click = "^7.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
imserv = 'imserv.__main__:runserver'
